// src/Types/product.interface.ts
export interface CardItem {
   brand: string;
   category: string;
   createdAt: string;
   description: string;
   dynamicFields: string | null;
   id: string;
   isActive: boolean;
   name: string;
   note: string | null;
   picture: string;
   price: number;
   sku: string;
   stock: number;
   updatedAt: string;
}